Moved non-common files outside of the common folder
