+++ firestarter/tests/test_dagger_workflow.py
@@ -1,5 +1,5 @@
-from firestarter.common.dagger_context import Context
-from firestarter.common.task_manager import TaskManager
+from firestarter.workflows.ci.dagger_context import Context
+from firestarter.workflows.ci.task_manager import TaskManager
 import anyio
 import os
 
@@ -9,9 +9,12 @@ def test_dagger_workflow():
     def myTestFunc(context):
         tm = TaskManager()
         tm.context = context
-        tm.load(os.path.join(
-            os.path.dirname(__file__), 'fixtures/test_workflow.yaml'
-        ))
+        tm.load(
+            os.path.join(
+                os.path.dirname(__file__), 'fixtures/test_workflow.yaml',
+            ),
+            '../workflows/ci/schema.json',
+        )
         tm.run()
 
     anyio.run(lambda: Context().start(myTestFunc))
