+++ firestarter/tests/test_workflow_validate.py
@@ -1,5 +1,5 @@
 from firestarter.common import validations
-from firestarter.common import task_manager
+from firestarter.workflows.ci import task_manager
 import os
 
 
@@ -8,11 +8,15 @@ class MockContext:
 
 def test_validate_schema():
     validations.validate_task_manager(
-        os.path.join(os.path.dirname(__file__), 'fixtures/workflow.yaml')
+        os.path.join(os.path.dirname(__file__), 'fixtures/workflow.yaml'),
+        '../workflows/ci/schema.json',
     )
 
 def test_validate_from_workflow():
     tm = task_manager.TaskManager()
     tm.context = MockContext()
-    tm.load(os.path.join(os.path.dirname(__file__), 'fixtures/workflow.yaml'))
+    tm.load(
+        os.path.join(os.path.dirname(__file__), 'fixtures/workflow.yaml'),
+        '../workflows/ci/schema.json',
+    )
