+++ firestarter/workflows/ci/__init__.py
@@ -1,5 +1,5 @@
 from .ci import CI
-from firestarter.common.dagger_context import Context
+from .dagger_context import Context
 
 def run(*, vars: dict, secrets: dict, config_file:str) -> None:
     context: Context = Context()
